# api/openapi.yaml
openapi: 3.1.0
info:
  title: Ivy API
  version: 0.0.1
  description: >
    The typed contract shared by the Go server and the SvelteKit client
    (STANDARDS.md section 6). Go and TypeScript code are generated from this
    file; committed outputs are drift-checked in CI. Only the read surface the
    frontend already calls is described so far; mutations and SSE event bodies
    are added per milestone. SSE event shapes are an extension because OpenAPI
    does not model them natively.
servers:
  - url: /api/v1
paths:
...
  /accounts/{id}:
    patch:
      operationId: updateAccountProfile
      summary: Rename an account or choose its icon
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountProfile"
      responses:
        "200":
          description: The updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          $ref: "#/components/responses/NotFound"
  /inbox:
    get:
      operationId: listInbox
      summary: Inbox summaries, newest first
      parameters:
        - name: account_id
          in: query
          schema:
            type: string
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        "200":
          description: A page of the inbox
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inbox"
  /messages/{id}:
    get:
      operationId: getMessage
      summary: One message with its sanitized body
      parameters:
        - $ref: "#/components/parameters/MessageID"
      responses:
        "200":
          description: The message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MailMessage"
        "404":
          $ref: "#/components/responses/NotFound"
  /messages/{id}/summary:
    get:
      operationId: getMessageSummary
      summary: The header alone, for a failed body
      parameters:
        - $ref: "#/components/parameters/MessageID"
      responses:
        "200":
          description: The summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MailSummary"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /events:
    get:
      operationId: streamEvents
      summary: Server-sent hints that something changed
      description: >
        A stream of small typed hints that say what to refetch (round 37). Each
        frame is `event: <type>` and one `data:` line holding an Event as JSON.
        There are no event ids and no replay: after any reconnect, refetch what
        you are showing. A slow client's queue is bounded and drops its oldest
        hints. Comment lines (`: keepalive`) are heartbeats, and the first frame
        is `retry: 3000`. At most 16 streams are open at once.
      responses:
        "200":
          description: An open event stream
          content:
            text/event-stream:
              schema:
                $ref: "#/components/schemas/Event"
        "503":
          description: Too many streams are open (`too_many_streams`) or Ivy is shutting down (`unavailable`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  parameters:
    MessageID:
      name: id
      in: path
      required: true
      schema:
        type: string
  responses:
    NotFound:
      description: No such resource
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable code the UI maps to its own copy
        message:
          type: string
        detail:
          type: string
    Event:
      type: object
      description: One hint on the event stream. It says what changed, never the new value.
      required: [type]
      properties:
        type:
          type: string
          enum: [message.changed, folder.changed, sync.state, outbox.state, health.alert]
          # Without explicit names the Go constant for sync.state is `SyncState`,
          # which collides with the SyncState type.
          x-enum-varnames:
            [EventTypeMessageChanged, EventTypeFolderChanged, EventTypeSyncState, EventTypeOutboxState, EventTypeHealthAlert]
        accountId:
          type: string
        folder:
          type: string
        code:
          type: string
          description: Stable reason, used by health.alert
...
    AccountSlot:
      type: integer
      enum: [1, 2, 3, 4, 5]
    SyncState:
      type: string
      enum: [ok, syncing, auth-failed, unreachable]
    Account:
      type: object
      required: [id, address, short, initial, slot, unread, smart, sync, syncNote, name, icon, photo]
      properties:
        id:
          type: string
        address:
          type: string
        short:
          type: string
        initial:
          type: string
        name:
          type: string
          description: The account's display name, empty when it has none
        icon:
          type: string
          description: A chosen emoji or glyph, empty when it has none
        photo:
          type: boolean
          description: Whether GET /accounts/{id}/photo serves a photo
        slot:
          $ref: "#/components/schemas/AccountSlot"
        unread:
          type: integer
        smart:
          type: boolean
        sync:
          $ref: "#/components/schemas/SyncState"
        syncNote:
          type: string
          description: >
            What the sync is doing, as a short phrase with no time in it ("Up to
            date", "Reading your mailbox, newest first"). The browser appends
            "synced 4 min ago" from syncedAt.
        syncedAt:
          type: string
          format: date-time
          description: >
            When the account's mailbox last synced, RFC 3339 in UTC. Absent until
            the first folder has synced.
        progress:
          type: number
          minimum: 0
          maximum: 1
    AccountProfile:
      type: object
      description: A partial rename; omitted fields keep their current value
      properties:
        displayName:
          type: string
          maxLength: 120
        icon:
          type: string
          maxLength: 16
    Attachment:
      type: object
      required: [id, name, size, kind]
      properties:
        id:
          type: string
        name:
          type: string
        size:
          type: string
        kind:
          type: string
          enum: [image, file]
        tone:
          type: string
          enum: [a, b]
        failed:
          type: boolean
    MailSummary:
      type: object
      required: [id, accountId, from, initials, date, subject, preview, unread, needs]
      properties:
        id:
          type: string
        accountId:
          type: string
        from:
          type: string
        initials:
          type: string
        date:
          type: string
          format: date-time
          description: >
            The message's instant, RFC 3339 in UTC. The browser formats it in the
            viewer's own zone and locale; a message with no Date header carries the
            zero instant (0001-01-01T00:00:00Z) and shows no time.
        subject:
          type: string
        preview:
          type: string
        unread:
          type: boolean
        needs:
          type: boolean
        tag:
          type: string
    MailMessage:
      type: object
      required:
        [id, accountId, from, initials, date, subject, preview, unread, needs,
         toShort, toFull, paragraphs, attachments]
      properties:
        id:
          type: string
        accountId:
          type: string
        from:
          type: string
        initials:
          type: string
        date:
          type: string
          format: date-time
          description: The message's instant, RFC 3339 in UTC (see MailSummary).
        subject:
          type: string
        preview:
          type: string
        unread:
          type: boolean
        needs:
          type: boolean
        tag:
          type: string
        toShort:
          type: string
        toFull:
          type: string
        summary:
          type: string
        html:
          type: string
          description: >-
            Server-sanitised HTML body, safe to place in the reader's sandboxed
            frame. Empty for a message with no HTML part; use paragraphs then.
        paragraphs:
          type: array
          items:
            type: string
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/Attachment"
    Inbox:
      type: object
      required: [items, needCount, unreadCount, readingWaiting]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/MailSummary"
        needCount:
          type: integer
        unreadCount:
          type: integer
        readingWaiting:
          type: integer
        nextCursor:
          type: string
...